課題提出及び履歴対応でデータベース構成変更

+++ src/main/java/jp/ac/ems/bean/QuestionBean.java
@@ -74,7 +74,7 @@ public class QuestionBean {
      * 問番(question number).
      */
     @Column(name = "number")
-    private Long number;
+    private Byte number;
 
     
     /**
